Fix a bug that leaks refresh loops when API errors occur
When an error occurs in the primary event handling loop, the cluster-agent
currently breaks out and creates a new client, but does not stop the current
client, which leaves the token refresh loop running. In the case of a
deterministic API error this can quickly lead to the buildup of thousands of
concurrent refresh loops.

This factors client instantiation out of "run" and into the reconnect loop,
and calls "stop" on clients before creating a new one.